Switch to gypsum backend for saving and loading datasets. (#22)
Similarly, we switch to the alabaster framework to store the data, rather than
RDS files. This should allow the files to be re-used in other languages.

Users can now grab SEs directly from the gypsum backend with fetchReference, and
they can perform listing and advanced searches directly on gypsum's SQLite
files. This lays the groundwork for removing dependencies on ExperimentHub.

All of the little per-dataset functions now call fetchReference to retrieve
content, where it is back-compatible to do so. We plan to soft-deprecate all
per-dataset functions to simplify maintenance; users should just call
fetchReference instead, and be responsible for any additional transformations.

We also provide instructions and functions for users to save and upload their
own datasets. This should simplify the process of making a contribution.